@@ -1,15 +1,14 @@
 {
   "name": "ng-select2",
-  "version": "1.0.8",
-  "description": "Angular 6 & 7 wrapper component of jQuery select2.",
+  "version": "1.1.0",
+  "description": "Angular 8 wrapper component of jQuery select2.",
   "repository": {
     "type": "git",
     "url": "https://github.com/tealpartners/ng-select2/"
   },
   "keywords": [
     "select2",
-    "angular4",
-    "angular7"
+    "angular8"
   ],
   "author": "Muhammad Arsal Abdullah <muhammadarsal.mail@gmail.com>",
   "contributors": [
@@ -18,14 +17,14 @@
   ],
   "license": "MIT",
   "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0 || ^7.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0 || ^7.0.0",
-    "@types/jquery": "^3.2.x",
+    "@angular/common": "^8.0.0",
+    "@angular/core": "^8.0.0",
+    "@types/jquery": "^3.4.x",
     "@types/select2": "^4.0.x",
     "select2": "^4.0.x"
   },
   "devDependencies": {
-    "@types/jquery": "^3.2.x",
+    "@types/jquery": "^3.3.x",
     "@types/select2": "^4.0.x",
     "select2": "^4.0.x"
   },
