@@ -2,24 +2,22 @@
   "name": "ng-select2",
   "version": "0.0.0",
   "dependencies": {
-    "@angular/animations": "7.0.1",
-    "@angular/common": "7.0.1",
-    "@angular/compiler": "7.0.1",
-    "@angular/core": "7.0.1",
-    "@angular/forms": "7.0.1",
-    "@angular/http": "7.0.1",
-    "@angular/platform-browser": "7.0.1",
-    "@angular/platform-browser-dynamic": "7.0.1",
-    "@angular/router": "7.0.1",
-    "@types/jquery": "3.3.22",
+    "@angular/animations": "8.0.1",
+    "@angular/common": "8.0.1",
+    "@angular/compiler": "8.0.1",
+    "@angular/core": "8.0.1",
+    "@angular/forms": "8.0.1",
+    "@angular/platform-browser": "8.0.1",
+    "@angular/platform-browser-dynamic": "8.0.1",
+    "@angular/router": "8.0.1",
+    "@types/jquery": "3.3.29",
     "@types/select2": "4.0.47",
-    "bootstrap": "4.1.3",
-    "core-js": "2.5.7",
-    "jquery": "3.3.1",
-    "popper.js": "1.14.4",
-    "rxjs": "6.3.3",
-    "select2": "4.0.6-rc.1",
-    "zone.js": "0.8.26"
+    "bootstrap": "4.3.1",
+    "jquery": "3.4.1",
+    "popper.js": "1.15.0",
+    "rxjs": "6.5.2",
+    "select2": "4.0.7",
+    "zone.js": "0.9.1"
   },
   "scripts": {
     "ng": "ng",
@@ -33,29 +31,18 @@
   },
   "private": true,
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.10.3",
-    "@angular-devkit/build-ng-packagr": "0.10.3",
-    "@angular/cli": "7.0.3",
-    "@angular/compiler-cli": "7.0.1",
-    "@angular/language-service": "7.0.1",
-    "@types/jasmine": "2.8.6",
-    "@types/jasminewd2": "2.0.3",
-    "@types/node": "8.9.4",
+    "@angular-devkit/build-angular": "0.800.3",
+    "@angular-devkit/build-ng-packagr": "0.800.3",
+    "@angular/cli": "8.0.3",
+    "@angular/compiler-cli": "8.0.1",
+    "@angular/language-service": "8.0.1",
+    "@types/node": "12.0.8",
     "angular-cli-ghpages": "0.5.3",
-    "codelyzer": "4.5.0",
-    "jasmine-core": "2.99.1",
-    "jasmine-spec-reporter": "4.2.1",
-    "karma": "1.7.1",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "1.4.2",
-    "karma-jasmine": "1.1.1",
-    "karma-jasmine-html-reporter": "0.2.2",
-    "ng-packagr": "4.4.0",
-    "protractor": "5.4.1",
+    "codelyzer": "5.1.0",
+    "ng-packagr": "5.3.0",
     "ts-node": "7.0.1",
-    "tsickle": "0.33.1",
     "tslib": "1.9.3",
-    "tslint": "5.11.0",
-    "typescript": "3.1.3"
+    "tslint": "5.17.0",
+    "typescript": "3.4.5"
   }
 }
